Deploy new version v2.1

+++ src/main/java/com/example/ratingbadge/model/Product.java
@@ -14,5 +14,5 @@ import javax.persistence.Table;
 public class Product extends BaseUUID {
     private String name;
     private String description;
-    private double avgStars;
+    private float avgStars;
 }
